test-norma-project/requirements.txt
norma-orm>=0.1.0
pydantic>=2.0.0
